It may be good to note once, that a == null actually a === null || a === undefined
I just found that out reading this code and dig into it further, but I posted it on Stack Overflow to see if other users know and it seemed like 50 out of 50 users didn't know that.

So if users read   `a == null`, they might actually feel it is just comparing to `null`.  A comment appearing once can help the situation.

It is interesting greenhorns use `==`, but one-step above greenhorn uses `===` in JavaScript, and then it is more pro, if somebody start to use `a == null` to compare it to `null` or `undefined`.